+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/handlers/workflow/DruidWorkflow.java
@@ -29,6 +29,7 @@ import com.yahoo.bard.webservice.web.handlers.VolatileDataRequestHandler;
 import com.yahoo.bard.webservice.web.handlers.WebServiceSelectorRequestHandler;
 import com.yahoo.bard.webservice.web.handlers.WeightCheckRequestHandler;
 import com.yahoo.bard.webservice.web.util.QueryWeightUtil;
+import com.yahoo.bard.webservice.web.util.QuerySignedCacheService;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -66,6 +67,7 @@ public class DruidWorkflow implements RequestWorkflowProvider {
     protected final @NotNull QuerySigningService<?> querySigningService;
     protected final @NotNull ObjectMapper mapper;
     protected final @NotNull VolatileIntervalsService volatileIntervalsService;
+    protected final @NotNull QuerySignedCacheService querySignedCacheService;
 
     /**
      * Constructor.
@@ -77,6 +79,7 @@ public class DruidWorkflow implements RequestWorkflowProvider {
      * @param partialDataHandler  Handler for dealing with the partial data step
      * @param querySigningService  Service to sign a query based on it's segment metadata
      * @param volatileIntervalsService  Service to get volatile intervals from
+     * @param querySignedCacheService The service for cache support
      * @param mapper  JSON mapper
      */
     @Inject
@@ -88,6 +91,7 @@ public class DruidWorkflow implements RequestWorkflowProvider {
             PartialDataHandler partialDataHandler,
             QuerySigningService<?> querySigningService,
             VolatileIntervalsService volatileIntervalsService,
+            QuerySignedCacheService querySignedCacheService,
             ObjectMapper mapper
     ) {
         this.dataCache = dataCache;
@@ -97,6 +101,7 @@ public class DruidWorkflow implements RequestWorkflowProvider {
         this.partialDataHandler = partialDataHandler;
         this.querySigningService = querySigningService;
         this.volatileIntervalsService = volatileIntervalsService;
+        this.querySignedCacheService = querySignedCacheService;
         this.mapper = mapper;
     }
 
@@ -114,7 +119,8 @@ public class DruidWorkflow implements RequestWorkflowProvider {
         if (CacheFeatureFlag.TTL.isOn()) {
             handler = new CacheRequestHandler(handler, dataCache, mapper);
         } else if (CacheFeatureFlag.LOCAL_SIGNATURE.isOn()) {
-            handler = new CacheV2RequestHandler(handler, dataCache, querySigningService, mapper);
+            handler =
+                    new CacheV2RequestHandler(handler, dataCache, querySigningService, querySignedCacheService, mapper);
         } else if (CacheFeatureFlag.ETAG.isOn()) {
             handler = new EtagCacheRequestHandler(
                     handler,
