+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/handlers/CacheWeightCheckRequestHandlerSpec.groovy
@@ -4,13 +4,8 @@ package com.yahoo.bard.webservice.web.handlers
 
 import static com.yahoo.bard.webservice.data.time.DefaultTimeGrain.DAY
 
-import static com.yahoo.bard.webservice.async.ResponseContextUtils.createResponseContext
-import static com.yahoo.bard.webservice.web.responseprocessors.ResponseContextKeys.MISSING_INTERVALS_CONTEXT_KEY
-
 import com.yahoo.bard.webservice.application.ObjectMappersSuite
 import com.yahoo.bard.webservice.druid.client.DruidWebService
-import com.yahoo.bard.webservice.druid.client.FailureCallback
-import com.yahoo.bard.webservice.druid.client.HttpErrorCallback
 import com.yahoo.bard.webservice.druid.client.SuccessCallback
 import com.yahoo.bard.webservice.druid.model.query.DruidAggregationQuery
 import com.yahoo.bard.webservice.druid.model.query.GroupByQuery
@@ -21,17 +16,13 @@ import com.yahoo.bard.webservice.metadata.QuerySigningService
 import com.yahoo.bard.webservice.metadata.SegmentIntervalsHashIdGenerator
 import com.yahoo.bard.webservice.web.apirequest.DataApiRequest
 import com.yahoo.bard.webservice.web.responseprocessors.ResponseProcessor
-import com.yahoo.bard.webservice.web.responseprocessors.ResponseContext
 import com.yahoo.bard.webservice.web.responseprocessors.WeightCheckResponseProcessor
 import com.yahoo.bard.webservice.web.util.QueryWeightUtil
 import com.yahoo.bard.webservice.web.RequestUtils
-import com.yahoo.bard.webservice.util.SimplifiedIntervalList
 import com.yahoo.bard.webservice.data.cache.TupleDataCache
 import com.yahoo.bard.webservice.data.cache.MemTupleDataCache
-import com.yahoo.bard.webservice.web.util.CacheService
+import com.yahoo.bard.webservice.web.util.QuerySignedCacheService
 
-import com.fasterxml.jackson.core.JsonFactory
-import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.core.JsonProcessingException
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.node.JsonNodeFactory
@@ -52,7 +43,7 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
     ObjectWriter writer
     TupleDataCache<String, Long, String> dataCache
     QuerySigningService<Long> querySigningService
-    CacheService cacheService
+    QuerySignedCacheService cacheService
     CacheWeightCheckRequestHandler handler
 
     RequestContext context
@@ -78,7 +69,7 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
         response = Mock(WeightCheckResponseProcessor)
         bardQueryInfo = BardQueryInfoUtils.initializeBardQueryInfo()
         json = new JsonNodeFactory().arrayNode()
-        cacheService = new CacheService(dataCache, querySigningService, MAPPER)
+        cacheService = new QuerySignedCacheService(dataCache, querySigningService, MAPPER)
         handler = new CacheWeightCheckRequestHandler(
                 next,
                 webService,
@@ -92,16 +83,16 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
         BardQueryInfoUtils.resetBardQueryInfo()
     }
 
-    def "Test constructor"() {
+    def "Check constructor"() {
         expect:
         handler.next == next
         handler.webService == webService
         handler.queryWeightUtil == queryWeightUtil
         handler.mapper == MAPPER
-        handler.cacheService == cacheService
+        handler.querySignedCacheService == cacheService
     }
 
-    def "Test handle request with request passing quick weight check"() {
+    def "Request passing quick weight check"() {
         setup:
         1 * queryWeightUtil.skipWeightCheckQuery(groupByQuery) >> true
         1 * next.handleRequest(context, request, groupByQuery, response) >> true
@@ -113,7 +104,7 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
         bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
     }
 
-    def "Test handleRequest without building callback"() {
+    def "Request handled without building callback"() {
         setup:
         final SuccessCallback success = Mock(SuccessCallback)
         groupByQuery.getGranularity() >> DAY
@@ -131,7 +122,7 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
             @Override
             public SuccessCallback buildCacheSuccessCallback(
                     final SuccessCallback classicSuccessCallback,
-                    final CacheService cacheService,
+                    final QuerySignedCacheService cacheService,
                     final DruidAggregationQuery<?> groupByQuery,
                     final ResponseProcessor response
             ) {
@@ -155,7 +146,7 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
         bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 1
     }
 
-    def "Test handleRequest without building callback with json error"() {
+    def "Request handled without building callback with json error"() {
         setup:
         final SuccessCallback success = Mock(SuccessCallback)
         groupByQuery.getGranularity() >> DAY
@@ -172,7 +163,7 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
             @Override
             public SuccessCallback buildCacheSuccessCallback(
                     final SuccessCallback classicSuccessCallback,
-                    final CacheService cacheService,
+                    final QuerySignedCacheService cacheService,
                     final DruidAggregationQuery<?> groupByQuery,
                     final ResponseProcessor response
             ) {
@@ -196,7 +187,7 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
         bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 1
     }
 
-    def "Test handle request on cache hit responds to the group by request"() {
+    def "On cache hit, handler responds to the group by request"() {
         setup:
         GroupByQuery groupByQuery = RequestUtils.buildGroupByQuery()
         QuerySigningService querySigningService = Mock(SegmentIntervalsHashIdGenerator)
@@ -207,7 +198,7 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
         RequestContext requestContext = new RequestContext(containerRequestContext, true)
         requestContext.isReadCache() >> true
         MAPPER.valueToTree(groupByQuery) >> json
-        cacheService = new CacheService(dataCache, querySigningService, MAPPER)
+        cacheService = new QuerySignedCacheService(dataCache, querySigningService, MAPPER)
         CacheWeightCheckRequestHandler handler = new CacheWeightCheckRequestHandler(
                 next,
                 webService,
@@ -232,7 +223,7 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
         bardQueryInfo.queryCounter.get(BardQueryInfo.FACT_QUERY_CACHE_HIT).get() == 1
     }
 
-    def "Test handle request on cache hit responds to the group by request with exception in cache read"() {
+    def "On cache hit, handler responds to the group by request with exception in cache read"() {
         setup:
         GroupByQuery groupByQuery = RequestUtils.buildGroupByQuery()
         QuerySigningService querySigningService = Mock(SegmentIntervalsHashIdGenerator)
@@ -242,7 +233,7 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
                 MultivaluedHashMap<String, String>)
         RequestContext requestContext = new RequestContext(containerRequestContext, true)
         requestContext.isReadCache() >> true
-        cacheService = new CacheService(dataCache, querySigningService, MAPPER)
+        cacheService = new QuerySignedCacheService(dataCache, querySigningService, MAPPER)
         CacheWeightCheckRequestHandler handler = new CacheWeightCheckRequestHandler(
                 next,
                 webService,
@@ -265,281 +256,5 @@ class CacheWeightCheckRequestHandlerSpec extends Specification {
 
         and: "The count of fact query cache hit is incremented by 1"
         bardQueryInfo.queryCounter.get(BardQueryInfo.FACT_QUERY_CACHE_HIT).get() == 1
-    }
-
-    def "Test cache hit with weight check"() {
-        setup:
-        SimplifiedIntervalList intervals = new SimplifiedIntervalList()
-        ResponseContext responseContext =  createResponseContext([(MISSING_INTERVALS_CONTEXT_KEY.name): intervals])
-        response.getResponseContext() >> responseContext
-        GroupByQuery groupByQuery = RequestUtils.buildGroupByQuery()
-        QuerySigningService querySigningService = Mock(SegmentIntervalsHashIdGenerator)
-        querySigningService.getSegmentSetId(_) >> Optional.of(1234L)
-        ContainerRequestContext containerRequestContext = Mock(ContainerRequestContext)
-        containerRequestContext.getHeaders() >> (["Bard-Testing": "###BYPASS###", "ClientId": "UI"] as
-                MultivaluedHashMap<String, String>)
-        RequestContext requestContext = new RequestContext(containerRequestContext, true)
-        cacheService = new CacheService(dataCache, querySigningService, MAPPER)
-        CacheWeightCheckRequestHandler handler = new CacheWeightCheckRequestHandler(
-                next,
-                webService,
-                queryWeightUtil,
-                MAPPER,
-                cacheService
-        )
-        requestContext.isReadCache() >> true
-        int expectedCount = 60
-        int limit = 100
-        SuccessCallback classicSuccessCallback = handler.buildSuccessCallback(context, request, groupByQuery, response, limit)
-        SuccessCallback success = handler.buildCacheSuccessCallback(classicSuccessCallback, cacheService, groupByQuery, response)
-        String weightResponse = """
-        [ {
-            "version" : "v1",
-            "timestamp" : "2012-01-01T00:00:00.000Z",
-            "event" : {
-                "count" : "$expectedCount"
-            }
-        } ]
-        """
-        JsonParser parser = new JsonFactory().createParser(weightResponse)
-        JsonNode jsonResult = MAPPER.readTree(parser)
-
-        expect:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-
-        when:
-        success.invoke(jsonResult)
-
-        then:
-        1 * next.handleRequest(context, request, groupByQuery, response)
-
-        and:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-    }
-
-    def "Test cache hit with weight check when exception thrown by cache read"() {
-        setup:
-        SimplifiedIntervalList intervals = new SimplifiedIntervalList()
-        ResponseContext responseContext =  createResponseContext([(MISSING_INTERVALS_CONTEXT_KEY.name): intervals])
-        response.getResponseContext() >> responseContext
-        GroupByQuery groupByQuery = RequestUtils.buildGroupByQuery()
-        QuerySigningService querySigningService = Mock(SegmentIntervalsHashIdGenerator)
-        querySigningService.getSegmentSetId(_) >> Optional.of(1234L)
-        ContainerRequestContext containerRequestContext = Mock(ContainerRequestContext)
-        containerRequestContext.getHeaders() >> (["Bard-Testing": "###BYPASS###", "ClientId": "UI"] as
-                MultivaluedHashMap<String, String>)
-        RequestContext requestContext = new RequestContext(containerRequestContext, true)
-        cacheService = new CacheService(dataCache, querySigningService, MAPPER)
-        CacheWeightCheckRequestHandler handler = new CacheWeightCheckRequestHandler(
-                next,
-                webService,
-                queryWeightUtil,
-                MAPPER,
-                cacheService
-        )
-        requestContext.isReadCache() >> true
-        int expectedCount = 60
-        int limit = 100
-        cacheService.readCache() >> {new Exception()}
-        SuccessCallback classicSuccessCallback = handler.buildSuccessCallback(context, request, groupByQuery, response, limit)
-        SuccessCallback success = handler.buildCacheSuccessCallback(classicSuccessCallback, cacheService, groupByQuery, response)
-        String weightResponse = """
-        [ {
-            "version" : "v1",
-            "timestamp" : "2012-01-01T00:00:00.000Z",
-            "event" : {
-                "count" : "$expectedCount"
-            }
-        } ]
-        """
-        JsonParser parser = new JsonFactory().createParser(weightResponse)
-        JsonNode jsonResult = MAPPER.readTree(parser)
-
-        expect:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-
-        when:
-        success.invoke(jsonResult)
-
-        then:
-        1 * next.handleRequest(context, request, groupByQuery, response)
-
-        and:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-    }
-
-
-    def "Test build and invoke cache success callback passes without cache hit"() {
-        setup:
-        SimplifiedIntervalList intervals = new SimplifiedIntervalList()
-        ResponseContext responseContext =  createResponseContext([(MISSING_INTERVALS_CONTEXT_KEY.name): intervals])
-        response.getResponseContext() >> responseContext
-        cacheService = new CacheService(dataCache, querySigningService, MAPPER)
-        CacheWeightCheckRequestHandler handler = new CacheWeightCheckRequestHandler(
-                next,
-                webService,
-                queryWeightUtil,
-                MAPPER,
-                cacheService
-        )
-        int expectedCount = 60
-        int limit = 100
-        SuccessCallback classicSuccessCallback = handler.buildSuccessCallback(context, request, groupByQuery, response, limit)
-        SuccessCallback success = handler.buildCacheSuccessCallback(classicSuccessCallback, cacheService, groupByQuery, response)
-        String weightResponse = """
-        [ {
-            "version" : "v1",
-            "timestamp" : "2012-01-01T00:00:00.000Z",
-            "event" : {
-                "count" : "$expectedCount"
-            }
-        } ]
-        """
-        JsonParser parser = new JsonFactory().createParser(weightResponse)
-        JsonNode jsonResult = MAPPER.readTree(parser)
-
-        expect:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-
-        when:
-        success.invoke(jsonResult)
-
-        then:
-        1 * next.handleRequest(context, request, groupByQuery, response)
-
-        and:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-    }
-
-    def "Test build and invoke cache success callback passes without cache hit with exception thrown by cache write"() {
-        setup:
-        SimplifiedIntervalList intervals = new SimplifiedIntervalList()
-        ResponseContext responseContext =  createResponseContext([(MISSING_INTERVALS_CONTEXT_KEY.name): intervals])
-        response.getResponseContext() >> responseContext
-        cacheService = new CacheService(dataCache, querySigningService, MAPPER)
-        CacheWeightCheckRequestHandler handler = new CacheWeightCheckRequestHandler(
-                next,
-                webService,
-                queryWeightUtil,
-                MAPPER,
-                cacheService
-        )
-        int expectedCount = 60
-        int limit = 100
-        SuccessCallback classicSuccessCallback = handler.buildSuccessCallback(context, request, groupByQuery, response, limit)
-        SuccessCallback success = handler.buildCacheSuccessCallback(classicSuccessCallback, cacheService, groupByQuery, response)
-        String weightResponse = """
-        [ {
-            "version" : "v1",
-            "timestamp" : "2012-01-01T00:00:00.000Z",
-            "event" : {
-                "count" : "$expectedCount"
-            }
-        } ]
-        """
-        JsonParser parser = new JsonFactory().createParser(weightResponse)
-        JsonNode jsonResult = MAPPER.readTree(parser)
-
-        expect:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-
-        when:
-        success.invoke(jsonResult)
-
-        then:
-        1 * next.handleRequest(context, request, groupByQuery, response)
-
-        and:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-    }
-
-    def "Test build and invoke cache success callback count too high"() {
-        setup:
-        SimplifiedIntervalList intervals = new SimplifiedIntervalList()
-        ResponseContext responseContext =  createResponseContext([(MISSING_INTERVALS_CONTEXT_KEY.name): intervals])
-        response.getResponseContext() >> responseContext
-        cacheService = new CacheService(dataCache, querySigningService, MAPPER)
-        CacheWeightCheckRequestHandler handler = new CacheWeightCheckRequestHandler(
-                next,
-                webService,
-                queryWeightUtil,
-                MAPPER,
-                cacheService
-        )
-        int expectedCount = 200
-        int limit = 100
-        SuccessCallback classicSuccessCallback = handler.buildSuccessCallback(context, request, groupByQuery, response, limit)
-        SuccessCallback success = handler.buildCacheSuccessCallback(classicSuccessCallback, cacheService, groupByQuery, response)
-        String weightResponse = """
-        [ {
-            "version" : "v1",
-            "timestamp" : "2012-01-01T00:00:00.000Z",
-            "event" : {
-                "count" : "$expectedCount"
-            }
-        } ]
-        """
-        JsonParser parser = new JsonFactory().createParser(weightResponse)
-        JsonNode jsonResult = new ObjectMapper().readTree(parser)
-        HttpErrorCallback ec = Mock(HttpErrorCallback)
-
-        expect:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-
-        when:
-        success.invoke(jsonResult)
-
-        then:
-        0 * next.handleRequest(context, request, groupByQuery, response)
-        1 * response.getErrorCallback(groupByQuery) >> ec
-        1 * ec.dispatch(507, _, _)
-
-        and:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-    }
-
-    def "Test build and invoke cache success callback invalid json"() {
-        setup:
-        SimplifiedIntervalList intervals = new SimplifiedIntervalList()
-        ResponseContext responseContext =  createResponseContext([(MISSING_INTERVALS_CONTEXT_KEY.name): intervals])
-        response.getResponseContext() >> responseContext
-        cacheService = new CacheService(dataCache, querySigningService, MAPPER)
-        CacheWeightCheckRequestHandler handler = new CacheWeightCheckRequestHandler(
-                next,
-                webService,
-                queryWeightUtil,
-                MAPPER,
-                cacheService
-        )
-        int expectedCount = 60
-        int limit = 100
-        SuccessCallback classicSuccessCallback = handler.buildSuccessCallback(context, request, groupByQuery, response, limit)
-        SuccessCallback success = handler.buildCacheSuccessCallback(classicSuccessCallback, cacheService, groupByQuery, response)
-        String weightResponse = """
-        [ {
-            "version" : "v1",
-            "timestamp" : "2012-01-01T00:00:00.000Z",
-            "event" : {
-                "count2" : "$expectedCount"
-            }
-        } ]
-        """
-        JsonParser parser = new JsonFactory().createParser(weightResponse)
-        JsonNode jsonResult = new ObjectMapper().readTree(parser)
-        FailureCallback fc = Mock(FailureCallback)
-
-        expect:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
-
-        when:
-        success.invoke(jsonResult)
-
-        then:
-        0 * next.handleRequest(context, request, groupByQuery, response)
-        1 * response.getFailureCallback(groupByQuery) >> fc
-        1 * fc.dispatch(_)
-
-        and:
-        bardQueryInfo.queryCounter.get(BardQueryInfo.WEIGHT_CHECK).get() == 0
     }
 }
